+++ packages/web3-types/CHANGELOG.md
@@ -161,4 +161,4 @@ Documentation:
 
 ### Added
 
--   add `asEip1193FullyCompatible` to `Web3BaseProvider` so every inherited class can have the returned value of `request` method, fully compatible with EIP-1193. (#6407)
+-   add `asEIP1193Provider` to `Web3BaseProvider` so every inherited class can have the returned value of `request` method, fully compatible with EIP-1193. (#6407)